requirements.txt
click==6.6
psutil==4.4.2
PyYAML==3.12